% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{bgcExecute}
\alias{bgcExecute}
\title{bgcExecute}
\usage{
bgcExecute(argv = "-a", iniFiles, simYearsOverride = -1, firstRun = TRUE)
}
\arguments{
\item{argv}{(Character Vector) : Arguments for the BiomeBGC library (same as 'bgc' commandline application)}

\item{iniFiles}{(String vector) : The list of ini files to simulate, one per site.
The files are expected to be specified for "go mode"}

\item{simYearsOverride}{The number of years to simulate (will override simyears from the ini file if specified).}

\item{firstRun}{this parameter will make the necessary changes in the ini file for the first run.
set to FALSE only when calling subsequent simulation steps in single step mode}
}
\value{
a list containing the resulting error code (0 means no error) and the ini file structure
}
\description{
Executes a simulation by calling the underlying BiomeBGC C library
}
