% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgcExecuteInternal}
\alias{bgcExecuteInternal}
\title{Execute BGC from R}
\usage{
bgcExecuteInternal(argv, iniFiles, simYearsOverride)
}
\arguments{
\item{argv}{commandline arguments to pass to bgc}

\item{iniFiles}{a vector of paths to the individual input \code{ini} files}

\item{simYearsOverride}{the number of years to use during simulation.  if <= 0, the field "number of simulation years" from the ini file is used}
}
\value{
an integer error code (0 means no error)
}
\description{
Execute BGC from R
}
\keyword{internal}
