% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binToText}
\alias{binToText}
\title{Converts single floating point binary data into ASCII text data}
\usage{
binToText(inDir, outDir, uniqueSuf, varNo = NA, years = NA, startYear = NA)
}
\arguments{
\item{inDir}{character string giving path to input data directory; \code{NA} means use current directory.}

\item{outDir}{character string giving path to output data directory; \code{NA} means use current directory.}

\item{uniqueSuf}{character string giving unique suffix.}

\item{varNo}{integer. the number of variables (columns); \code{NA} means all columns.}

\item{years}{integer. the number of years of the data (rows); \code{NA} means all rows.}

\item{startYear}{integer. the starting year of the data; \code{NA} means the first row.}
}
\value{
invoked for side effect of converting data files
}
\description{
Converts single floating point binary data into ASCII text data
}
