% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomeBGCR-package.R
\docType{package}
\name{BiomeBGCR-package}
\alias{BiomeBGCR}
\alias{BiomeBGCR-package}
\title{BiomeBGCR: Run Biome-BGC models from R}
\description{
Call and run Biome-BGC ecophysiological models in R.
}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Jean-Francois Lavoie \email{atmailjfl@gmail.com}
  \item Celine Boisvenue \email{celine.boisvenue@nrcan-rncan.gc.ca}
}

Other contributors:
\itemize{
  \item Numerical Terradynamics Simulation Group (NTSG) [copyright holder]
}

}
\keyword{internal}
