% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareASCIIFiles}
\alias{compareASCIIFiles}
\title{Compares two ASCII files}
\usage{
compareASCIIFiles(fileNameA, fileNameB)
}
\arguments{
\item{fileNameA}{the name of file A}

\item{fileNameB}{the name of file B}
}
\value{
TRUE if files are identical, false otherwise
}
\description{
Compares two ASCII files
}
