% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareASCIILines}
\alias{compareASCIILines}
\title{compareASCIILines}
\usage{
compareASCIILines(fileNameA, fileNameB, lineIndexA, lineIndexB)
}
\arguments{
\item{fileNameA}{the name of file A}

\item{fileNameB}{the name of file B}

\item{lineIndexA}{the line number to compare in file A}

\item{lineIndexB}{the line number to compare in file B}
}
\value{
TRUE if lines are identical, false otherwise
}
\description{
Compares a specific line in two ASCII files
}
