% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniFindSection}
\alias{iniFindSection}
\title{iniFindSection}
\usage{
iniFindSection(con, tag)
}
\arguments{
\item{con}{the connection to the file to scan from}

\item{tag}{The name of the section to look for into the ini file}
}
\value{
the line of the ini file starting with the specified tag
}
\description{
This function scans lines until it finds the specified section tag
}
