% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniGet}
\alias{iniGet}
\title{Get accessor for the BiomeBGC ini data structure}
\usage{
iniGet(ini, section, index)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}

\item{section}{the section tag (MET_INPUT, RESTART, etc..)}

\item{index}{the index (or indices) of the parameters.}
}
\value{
the value of the field (character vector)
}
\description{
Get accessor for the BiomeBGC ini data structure
}
\seealso{
iniRead
}
