% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniMakeSingleStep}
\alias{iniMakeSingleStep}
\title{iniMakeSingleStep}
\usage{
iniMakeSingleStep(ini, firstRun)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}

\item{firstRun}{set to TRUE if this ini file is the first one to be executed after spinup, FALSE otherwise}
}
\description{
Changes the ini file to enable single step mode
}
