% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniMakeSpinup}
\alias{iniMakeSpinup}
\title{iniMakeSpinup}
\usage{
iniMakeSpinup(ini)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}
}
\description{
Changes the ini file to that it will execute as a spinup simulation
}
