% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniParseSectionLine}
\alias{iniParseSectionLine}
\title{iniParseSectionLine}
\usage{
iniParseSectionLine(line, lookForSecondField = TRUE)
}
\arguments{
\item{line}{the line to parse into}

\item{lookForSecondField}{specifies if this line should contain 3 (TRUE) or 2 (FALSE) columns}
}
\value{
a data frame containing three string fields (value, unit, comment)
}
\description{
This function parses a section line (string) and returns a data frame for this line
}
