% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniRead}
\alias{iniRead}
\title{iniRead}
\usage{
iniRead(fileName)
}
\arguments{
\item{fileName}{the name of the file to open}
}
\value{
a list of data frames containing one dataframe per section
}
\description{
Reads the BGC-Biome ini file
}
