% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadClimChange}
\alias{iniReadClimChange}
\title{iniReadClimChange}
\usage{
iniReadClimChange(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the CLIM_CHANGE ini section
}
