% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadCO2Control}
\alias{iniReadCO2Control}
\title{iniReadCO2Control}
\usage{
iniReadCO2Control(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the CO2_CONTROL ini section
}
