% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadOutput}
\alias{iniReadOutput}
\title{iniReadOutput}
\usage{
iniReadOutput(con, section)
}
\arguments{
\item{con}{the connection to the file to scan from}

\item{section}{the tag of the section to read (DAILY_OUTPUT or ANNUAL_OUTPUT)}
}
\value{
the data frame containing the section data
}
\description{
This function can be used to read either DAILY_OUTPUT or ANNUAL_OUTPUT sections
}
