% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadOutputControl}
\alias{iniReadOutputControl}
\title{iniReadOutputControl}
\usage{
iniReadOutputControl(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the OUTPUT_CONTROL ini section
}
