% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadRampNdep}
\alias{iniReadRampNdep}
\title{iniReadRampNdep}
\usage{
iniReadRampNdep(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the RAMP_NDEP ini section
}
