% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadSite}
\alias{iniReadSite}
\title{iniReadSite}
\usage{
iniReadSite(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the SITE ini section
}
