% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniReadWState}
\alias{iniReadWState}
\title{iniReadWState}
\usage{
iniReadWState(con)
}
\arguments{
\item{con}{the connection to the file to scan from}
}
\value{
the data frame containing the section data
}
\description{
This function reads the W_STATE ini section
}
