% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniSet}
\alias{iniSet}
\title{iniSet}
\usage{
iniSet(ini, section, index, value)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}

\item{section}{the section tag (MET_INPUT, RESTART, etc..)}

\item{index}{the index (or indices) of the parameters.}

\item{value}{the value (character vector) to set the value to}
}
\value{
the updated ini data
}
\description{
Set accessor for the BiomeBGC ini data structure
}
\seealso{
iniRead
}
