% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniWrite}
\alias{iniWrite}
\title{Writes the ini data to an INI file}
\usage{
iniWrite(ini, fileName, title = "", description = "")
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}

\item{fileName}{(character vector) the name of the file to write to}

\item{title}{(character vector) the title of the file (will be written in the first line)}

\item{description}{(character vector) the description of the file (will be written in the first line)}
}
\description{
Writes the ini data to an INI file
}
\seealso{
iniRead
}
