% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bgcExecute}
\alias{bgcExecute}
\title{Execute BGC from R}
\usage{
bgcExecute(argv, iniFiles)
}
\arguments{
\item{`iniFiles`}{is the path to the input \code{ini} files}
}
\value{
an integer error code (0 means no error)
}
\description{
Execute BGC from R
}
