% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{bgcExecuteSpinup}
\alias{bgcExecuteSpinup}
\title{bgcExecuteSpinup}
\usage{
bgcExecuteSpinup(argv = "-a", iniFiles)
}
\arguments{
\item{argv}{(Character Vector) : Arguments for the BiomeBGC library (same as 'bgc' commandline application)}

\item{iniFiles}{(String vector) : The list of ini files to simulate, one per site.
The files can be either spinup ini files or "go" ini files.  A "go" file will be automatically modified for spinup phase.}
}
\value{
a list containing the resulting error code (0 means no error) and the ini file structure
}
\description{
Executes a spinup simulation by calling the underlying BiomeBGC C library
}
