% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniFixPaths}
\alias{iniFixPaths}
\title{iniFixPaths}
\usage{
iniFixPaths(ini)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}
}
\description{
Fixes all the relative paths inside the ini file so that the C++ lib can use the resulting the correct absolute paths.
}
