% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniFile.R
\name{iniFixPaths}
\alias{iniFixPaths}
\title{Fix all the relative paths inside the ini file}
\usage{
iniFixPaths(ini)
}
\arguments{
\item{ini}{the ini data structure returned by iniRead()}
}
\description{
C++ lib requires absolute paths.
}
